requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
importlib-metadata==6.7.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.2.0
pytest==7.4.0
python-dotenv==1.0.0
SQLAlchemy==2.0.17
tomli==2.0.1
typing_extensions==4.6.3
Werkzeug==2.3.6
zipp==3.15.0
